fully unify groups repository
